weatherbotapi/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
Markdown==3.3.4
django-filter==2.4.0
psycopg2-binary==2.8.5
requests